# web/swagger/openapi.yaml
openapi: 3.1.0
info:
  title: WMS Test API
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development
paths:
  /auth/login:
    post:
      summary: Login with email and password
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Successful authentication
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/refresh:
    post:
      summary: Refresh access token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: New access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshResponse'
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /items:
    get:
      summary: List items
      operationId: listItems
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: q
          schema:
            type: string
          description: Filter by SKU or name (case insensitive)
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number (1-based)
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Page size
        - in: query
          name: sort
          schema:
            type: string
            enum:
              - name
              - sku
              - qty
              - updated_at
            default: name
          description: Sort field
        - in: query
          name: dir
          schema:
            type: string
            enum:
              - asc
              - desc
            default: asc
          description: Sort direction
      responses:
        '200':
          description: A paginated list of items
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemsResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /items/{id}:
    get:
      summary: Get item by ID
      operationId: getItem
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: Item identifier
      responses:
        '200':
          description: Item found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '404':
          description: Item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      summary: Adjust quantity by delta
      operationId: patchItemQuantity
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: Item identifier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchItemRequest'
      responses:
        '200':
          description: Updated item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '400':
          description: Invalid request or negative quantity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: test@example.com
        password:
          type: string
          example: Password123!
    LoginResponse:
      type: object
      required:
        - access_token
        - refresh_token
      properties:
        access_token:
          type: string
          description: JWT access token
        refresh_token:
          type: string
          description: Refresh token identifier
    RefreshRequest:
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
          description: Refresh token identifier
    RefreshResponse:
      type: object
      required:
        - access_token
      properties:
        access_token:
          type: string
          description: New access token
    PatchItemRequest:
      type: object
      required:
        - qtyDelta
      properties:
        qtyDelta:
          type: number
          description: Delta to apply to inventory quantity (negative allowed)
          example: -2
    Item:
      type: object
      required:
        - id
        - sku
        - name
        - qty
        - location
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        sku:
          type: string
        name:
          type: string
        qty:
          type: integer
        location:
          type: string
        updated_at:
          type: string
          format: date-time
    ItemsResponse:
      type: object
      required:
        - items
        - page
        - page_size
        - total
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Item'
        page:
          type: integer
        page_size:
          type: integer
        total:
          type: integer
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          enum:
            - unauthorized
            - not_found
            - bad_request
            - conflict
            - internal
        message:
          type: string